run action in foreground instead
I realized there's no reason to background the action, and it just
makes for a race condition with the finished_test, as the test could
kick off and complete before the action has even fully started. This
changes it to just run the action in the foreground, but still do the
finished_test afterward. In most cases, if the action command
finishes, that should be enough, but I'll leave the test just in
case. It's more flexible this way. The default of `true` for the
finished test means you can just ignore it if you don't need it.